exam dan result blm fix
